backend/requirements.txt
asgiref==3.2.10
Django==3.2.10
djangorestframework==3.12.4
Pillow==8.0.1
python-dotenv==1.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1 